refactor(backend): split tests of settings in sub-parts (#5312)
# Motivation

Just to keep it more organized, we split the `settings` tests into the
ones related to the `networks` and the ones related to the `dapps`.
