+++ src/backend/tests/it/settings/mod.rs
@@ -1,0 +1,2 @@
+mod dapp_settings;
+mod networks_settings;
